requirements.txt
torch==1.9.0
matplotlib==3.3.4
pandas==1.2.5
scipy==1.7.1
torchvision==0.10.0
numpy==1.21.2
seaborn==0.11.1
imageio==2.9.0
natsort==7.1.1
Pillow==8.4.0
tensorflow==2.6.0
tqdm==4.62.3